Inline ResizeObserver to avoid unwanted compilation
