--- src/ts/greedy-nav/__mocks__/ResizeObserver.ts
@@ -1,22 +1,0 @@
-/* eslint-disable class-methods-use-this */
-class ResizeObserver {
-  callback: ResizeObserverCallback;
-
-  constructor(callback: ResizeObserverCallback) {
-    this.callback = callback;
-  }
-
-  observe(target: Element, options?: ResizeObserverOptions | undefined): void {
-    // do nothing
-  }
-
-  unobserve(): void {
-    // do nothing
-  }
-
-  disconnect(): void {
-    // do nothing
-  }
-}
-
-export default ResizeObserver;
